Trait rusty_trees::radix_trie::KeyRef

pub trait KeyRef<Key: ?Sized>: ToOwned<Owned = Key> {
    fn is_empty(&self) -> bool;
    fn prefix<'a>(a: &'a Self, b: &'a Self) -> (&'a Self, &'a Self, &'a Self);
    fn concat(iter: &mut dyn Iterator<Item = &Self>) -> Key;
}

Trait that the key’s Ref type must implement. All operations are performed on references to the keys within the trie, and not the owned values themselves.

Required Methods

fn is_empty(&self) -> bool

Is the key empty?

fn prefix<'a>(a: &'a Self, b: &'a Self) -> (&'a Self, &'a Self, &'a Self)

For keys a and b, return the common prefix between the two, and the remaining parts of the keys that remain

fn concat(iter: &mut dyn Iterator<Item = &Self>) -> Key

Concatenate an iterator of Self into the owned type